novahacks/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
